api/requirements.txt
flask==1.1.2
requests==2.24.0
python-dotenv==0.14.0